// src/accounts/dto/accounts.address.update.dto.ts
export class AccountsAddressUpdateDto {
  place_name: string;
  recipient: string;
  address_number: string;
  moo_soi_road: string;
  province: string;
  district: string;
  subdistrict: string;
  state: string;
  city: string;
  area: string;
  postal_code: string;
  contact_number: string;
  note_to_delivery: string;
}